Creating configDisplayer lib
